requirements.txt
torch
torchvision
Pillow
scikit-learn
py-cpuinfo